requirements.txt
Django>=4.1.6,<4.2
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.5,<3.0
drf-spectacular>=0.25.1,<0.26